+++ src/main/java/org/mariadb/r2dbc/client/ClientImpl.java
@@ -16,7 +16,6 @@
 
 package org.mariadb.r2dbc.client;
 
-import io.netty.channel.ChannelOption;
 import io.r2dbc.spi.R2dbcNonTransientResourceException;
 import java.net.SocketAddress;
 import java.util.Queue;
@@ -47,12 +46,7 @@ public final class ClientImpl extends ClientBase {
       MariadbConnectionConfiguration configuration) {
 
     TcpClient tcpClient = TcpClient.create(connectionProvider).remoteAddress(() -> socketAddress);
-    if (configuration.getConnectTimeout() != null) {
-      tcpClient =
-          tcpClient.option(
-              ChannelOption.CONNECT_TIMEOUT_MILLIS,
-              Math.toIntExact(configuration.getConnectTimeout().toMillis()));
-    }
+    tcpClient = setSocketOption(configuration, tcpClient);
     return tcpClient.connect().flatMap(it -> Mono.just(new ClientImpl(it, configuration)));
   }