+++ CHANGELOG.md
@@ -1,4 +1,15 @@
 # Change Log
+## [0.8.4](https://github.com/mariadb-corporation/mariadb-connector-r2dbc/tree/0.8.4) (29 Sep 2020)
+[Full Changelog](https://github.com/mariadb-corporation/mariadb-connector-r2dbc/compare/0.8.3...0.8.4)
+
+First Release Candidate release.
+
+Changes compared to 0.8.3.beta1:
+[R2DBC-9] Non pipelining prepare is close 2 time when race condition cache
+[R2DBC-8] synchronous close
+[R2DBC-7] authentication error when using multiple classloader
+[R2DBC-6] socket option configuration addition for socket Timeout, keep alive and force RST
+
 ## [0.8.3](https://github.com/mariadb-corporation/mariadb-connector-r2dbc/tree/0.8.3) (23 Jul 2020)
 [Full Changelog](https://github.com/mariadb-corporation/mariadb-connector-r2dbc/compare/0.8.2...0.8.3)